vt(4): Test if the cursor is shown only once
Later, we just see if the "struct mouse_cursor" pointer is set. This
avoids the need to mess with all the conditions several times; this has
been error prone.

While here, rename the variable "m" to a more meaningful "cursor", like
it's done elsewhere in the code.

MFC after:	1 week


git-svn-id: https://svn.freebsd.org/base/head@270290 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f