Unlock the right lock.
The adist_remote_lock is not held in this place, whereas the
adist_recv_list_lock lock is and is picked up during the next iteration.

I found this by annotating our libpthread with Clang's -Wthread-safety
attributes. I will send out a patch for this in the nearby future,
because it's awesome.

MFC after:	2 weeks


git-svn-id: https://svn.freebsd.org/base/head@270749 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f