Fix a tab that somehow became 8 spaces.
Remove parentheses in a return statement to be consistent with the
rest of the file.

Rename sin_pi() in the float version to sin_pif().

Remove large comment that precedes sin_pif().  The comment
duplicates a comment in e_lgamma_r.c where the algorithm
is documented.

Requested by:	bde


git-svn-id: https://svn.freebsd.org/base/head@270932 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f